requirements.txt
django
djangorestframework
groq
psycopg2-binary
pgvector
google-generativeai
openai
django-environ